backend/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-dotenv==1.0.0
dj-database-url==3.0.1
psycopg2-binary==2.9.9
openai==1.54.5
requests==2.32.4
Pillow==11.0.0
django-filter==24.3